Use DynamicTestExecutor to fix incompatibility
Prior to this commit the dynamically discovered tests were reported to
the `EngineExecutionListener` directly which caused an incompatibility
with JUnit Platform 1.3.

This commit changes the runtime behavior. Previously, a scope block
was executed entirely before executing the first test. Now, the tests
are executed immediately when they are encountered.

Fixes #462.